use pdfdocencoding when parsing strings
