+++ src/UglyToad.PdfPig.Tests/Parser/Parts/FileHeaderParserTests.cs
@@ -152,7 +152,7 @@ three %PDF-1.6";
 
             var bytes = new ByteArrayInputBytes(input);
 
-            var scanner = new CoreTokenScanner(bytes, ScannerScope.None);
+            var scanner = new CoreTokenScanner(bytes, true, ScannerScope.None);
 
             var result = FileHeaderParser.Parse(scanner, bytes, false, log);
