+++ src/UglyToad.PdfPig.Tokenization/DictionaryTokenizer.cs
@@ -7,6 +7,7 @@
 
     internal class DictionaryTokenizer : ITokenizer
     {
+        private readonly bool usePdfDocEncoding;
         private readonly IReadOnlyList<NameToken> requiredKeys;
 
         public bool ReadsNextByte { get; } = false;
@@ -14,12 +15,16 @@
         /// <summary>
         /// Create a new <see cref="DictionaryTokenizer"/>.
         /// </summary>
+        /// <param name="usePdfDocEncoding">
+        /// Whether to read strings using the PdfDocEncoding.
+        /// </param>
         /// <param name="requiredKeys">
         /// Can be provided to recover from errors with missing dictionary end symbols if the
         /// set of keys expected in the dictionary are known.
         /// </param>
-        public DictionaryTokenizer(IReadOnlyList<NameToken> requiredKeys = null)
+        public DictionaryTokenizer(bool usePdfDocEncoding, IReadOnlyList<NameToken> requiredKeys = null)
         {
+            this.usePdfDocEncoding = usePdfDocEncoding;
             this.requiredKeys = requiredKeys;
         }
 
@@ -75,7 +80,7 @@
                 return false;
             }
 
-            var coreScanner = new CoreTokenScanner(inputBytes, ScannerScope.Dictionary);
+            var coreScanner = new CoreTokenScanner(inputBytes, usePdfDocEncoding, ScannerScope.Dictionary);
 
             var tokens = new List<IToken>();
